refactor(web): consolidate duplicated plugin link-info hooks
The dropbox, google, synology-drive and one-drive plugins each shipped a
byte-identical `useLinkInfo` that derives a `ID: <fileId>` label, and the
server and webdav plugins shipped an identical `filePath`-based one. Move
both implementations into `plugins/common/linkInfo.ts` as
`useFileIdLinkInfo` and `useFilePathLinkInfo`, and point each plugin at the
shared export. Behavior is unchanged; a plugin can still diverge later by
providing its own implementation.
