+++ apps/web/src/plugins/one-drive/index.tsx
@@ -10,7 +10,7 @@ import { Provider } from "./Provider"
 import { InfoScreen } from "./InfoScreen"
 import { UploadBook } from "./UploadBook"
 import { UpsertFile } from "./UpsertFile"
-import { useLinkInfo } from "./useLinkInfo"
+import { useFileIdLinkInfo as useLinkInfo } from "../common/linkInfo"
 import { useRefreshMetadata } from "./useRefreshMetadata"
 import { useSynchronize } from "./useSynchronize"
 import { useSyncSourceInfo } from "./useSyncSourceInfo"
