+++ apps/web/src/plugins/server/index.tsx
@@ -6,7 +6,7 @@ import { DataSourceForm } from "./DataSourceForm"
 import { DownloadBook } from "./DownloadBook"
 import { InfoScreen } from "./InfoScreen"
 import { UploadBook } from "./UploadBook"
-import { useLinkInfo } from "./useLinkInfo"
+import { useFilePathLinkInfo as useLinkInfo } from "../common/linkInfo"
 import { useRefreshMetadata } from "./useRefreshMetadata"
 import { useSyncSourceInfo } from "./useSyncSourceInfo"
 import { useSynchronize } from "./useSynchronize"
